completed with routes

+++ backend/controller/tokens.controller.js
@@ -1,8 +1,12 @@
 import { request, response } from 'express';
-import {PrismaClient} from '../generated/prisma/index.js'
-const prisma = new PrismaClient();
+import initPrisma from '../dbinit.js';
+
+const prisma = await initPrisma()
 export const tokensAdd = async (req = request , res = response) => {
     try{
+        if (req.user.role !== 'admin') {
+            return res.status(403).json({ message: 'Access denied: Admins only' });
+        }
         let {departmentId,patient_name, status } = req.body
         console.log(req.body)
         
@@ -27,6 +31,9 @@ export const tokensAdd = async (req = request , res = response) => {
 
 export const tokensDelete = async (req = request , res = response) => {
     try{
+        if (req.user.role !== 'admin') {
+            return res.status(403).json({ message: 'Access denied: Admins only' });
+        }
         let {token_id,departmentId,status } = req.body
         await prisma.token_queue.deleteMany({
             where: {
@@ -46,6 +53,9 @@ export const tokensDelete = async (req = request , res = response) => {
 
 export const tokensGet = async(req = request , res = response) => {
     try{
+        if (req.user.role !== 'admin') {
+            return res.status(403).json({ message: 'Access denied: Admins only' });
+        }
        let filters = []
         if(req.body){
             let { token_id, departmentId, status } = req.body;
@@ -55,15 +65,36 @@ export const tokensGet = async(req = request , res = response) => {
              status ? { status } : undefined
            ].filter(Boolean);
         }
-
-
-let data = await prisma.token_queue.findMany(
-  filters.length > 0
-    ? { where: { OR: filters } }
-    : {} 
-);
+        let data = await prisma.token_queue.findMany(
+            filters.length > 0
+                ? { where: { OR: filters } }
+                : {} 
+        );
 
         res.status(200).json(data)
+    }catch(e){
+        console.log(e)
+        res.status(500).json({message:"error"})
+    }
+}
+
+export const tokensUpdate = async (req = request , res = response) => {
+    try{
+        if (req.user.role !== 'admin') {
+            return res.status(403).json({ message: 'Access denied: Admins only' });
+        }
+        let {token_id,departmentId, patient_name, status } = req.body
+        await prisma.token_queue.update({
+            where: {
+                token_id 
+            },
+            data: {
+                departmentId,
+                patient_name,
+                status
+            }
+        });
+        res.status(200).json({message:"success"})
     }catch(e){
         console.log(e)
         res.status(500).json({message:"error"})
